pkg/cvo: Render status.Failure in ReconciliationIssues
Walk the error's tree using Go's [1]:

  Unwrap() error
  Unwrap() []error

APIs.  Also include Errors(), because that's what our
k8s.io/apimachinery implementation currently uses:

  $ git --no-pager grep '^func.* [A-Z].*) \[]error {' -- vendor
  vendor/k8s.io/apimachinery/pkg/runtime/error.go:func (e *strictDecodingError) Errors() []error {
  vendor/k8s.io/apimachinery/pkg/runtime/helper.go:func DecodeList(objects []Object, decoders ...Decoder) []error {
  vendor/k8s.io/apimachinery/pkg/util/errors/errors.go:func (agg aggregate) Errors() []error {
  vendor/k8s.io/client-go/tools/clientcmd/validation.go:func (e errConfigurationInvalid) Errors() []error {

I'm stable-sorting the children so consumers don't have to worry about
confusing users by flipping things around depending on which worker
goroutine happened to bump into trouble with a manifest first.

I'm using the meta.v1 length limit [2], even though that technically
doesn't apply to ClusterOperatorStatusCondition [3], because it's
unlikely that the signal-to-noise in such a long message is high
enough for users to be able to find actionable information.  And
failing to write the ClusterVersion resource because we've overrun an
object-size cap wouldn't be much help to users either.

[1]: https://pkg.go.dev/errors#pkg-overview
[2]: https://github.com/kubernetes/apimachinery/blob/e696ec55a32e2177cd5f2fe1ded91d9485aa5c64/pkg/apis/meta/v1/types.go#L1575-L1577
[3]: https://github.com/openshift/api/blob/1e963d8dc4663f4f004f44fd58459381a771bdb5/config/v1/types_cluster_operator.go#L149-L153