pkg/cvo/cvo_scenarios_test.go: Modify TestCVO_ParallelError
Due to the introduced filtering of UpdateError errors before setting the
Failing condition, it is needed to update the TestCVO_ParallelError
test, as its errors are getting rightfully filtered due to their
UpdateEffect being None. This commit is utilizing this chance to
update the UpdateEffect of one of the errors to test the filtering here
as well.