OTA-1269: Add scaffolding of the Update Status Controller
As agreed in OTA-1269, the controller uses ConfigMap as a pseudo-api
until we have a real API committed to openshift/api.

The Update Status Controller is a new Cobra subcommand of the CVO
binary. The new command launches two controllers for now:

1. **Control Plane Informer** (CPI) monitors ClusterVersion resources
(for now) and produces a respective ClusterVersion status insights.
The functionality is limited, because part of the logic necessary
to populate all fields on a CV insight requires watching also
ClusterOperator monitoring and processing, which is not present yet.

1. **Update Status Controller** (USC) receives insights from the CPI and
maintains the ConfigMap that mimicks the eventual Status API.