OTA-1269: Add scaffolding of the Update Status Controller
As agreed in OTA-1269, the controller uses ConfigMap as a pseudo-api
until we have a real API committed to openshift/api.

The Update Status Controller is a new Cobra subcommand of the CVO
binary. The new command launches two controllers for now:

1. **Control Plane Informer** (CPI) monitors ClusterVersion resources
(for now) and produces a respective ClusterVersion status insights.
The functionality is limited, because part of the logic necessary
to populate all fields on a CV insight requires watching also
ClusterOperator monitoring and processing, which is not present yet.

1. **Update Status Controller** (USC) receives insights from the CPI and
maintains the ConfigMap that mimicks the eventual Status API.

Squashed review-feedback commits:

USC: Treat version specifically in USC

library-go code wants a different notion of version (k8s `version.Info` struct) than CVO uses. To avoid messing with CVO at this stage of USC, make USC use a minimal necessary version code, isolated from existing CVO. Mark this as tech debt to be solved before we graduate to TechPreview

USC: Log messages sent by CPI at V(4)

USC: Explain not waiting for informer cache sync

USC: Start insight receiver as PostStartHook

Previously the code spawned the receiver goroutine in the constructor, which was surprising and poorly structured. Spawn the receiver in a `PostStartHook` that is called by the library code after the controller is started. This also has an advantage of having access to controller queue, which means the receiver can trigger standard controller sync instead of calling controller internals.

Additionally, clearly document the communication mechanism as a scaffold code to be replaced or improved if needed