OTA-1269: Add `DevPreviewNoUpgrade`-gated manifests
- Namespace
- ServiceAccount
- Role that allows read, watch & update ConfigMaps (+binding)
- ClusterRole that allows read & watch ConfigVersions (+rolebinding)
- Deployment

Squashed review commits:

USC: Do not tolerate taints

The original Deployment was copied from CVO which contained a bunch of
strong tolerations. It does not seem USC needs to be this robust against
node conditions and can be easily evicted if necessary.

https://github.com/openshift/enhancements/blob/master/CONVENTIONS.md#taints-and-tolerations

USC: Use `openshift-user-critical` priority class

The original USC deployment was copied from CVO which is obviously very
important component and should be protected from OOM kills and preemption
but USC does not seem to need it.

https://github.com/openshift/enhancements/blob/master/CONVENTIONS.md#priority-classes

> If it is fine for your operator/operand to be preempted by user-workload and OOMKilled use openshift-user-critical priority class