WIP: pkg/cvo/updatepayload: Drop the Job controller for release-manifests downloads
Previously, the CVO launched a Job and waited for it to complete to
get manifests for an incoming release payload.  But the Job controller
doesn't bubble up details about why the pod has trouble
(e.g. Init:SignatureValidationFailed), so to get those details, we
need direct access to the Pod.  The Job controller doesn't seem like
it's adding much value here, so we're dropping it and monitoring the
Pod ourselves.