OTA-861: Set Upgradeable=False when there is an upgrade in progress (#1080)
* inhibit the 2nd minor version upgrade

* Block Y stream upgrade if any upgrade is in progress

This commits entends the guard on the 2nd Y-stream upgrade,
i.e., blocking an Y-stream upgrade if there is already an
Y-stream upgrade in progress,
to the guard on any Y-stream upgrade if there is already an
upgrade in progress, reguardless of Y-stream or Z-stream.

For example, it covers the case 4.14.15-> 4.14.35 -> 4.15.29
where the upgrade 4.14.35 -> 4.15.29 is blocked until the upgrade
4.14.15-> 4.14.35 completes.

Note that we still allow for upgrade to 4.y+1.z''
in the middle of upgrade 4.y.z -> 4.y+1.z', even though direct upgrade
4.y.z -> 4.y+1.z'' might not be supported.
This is because the ugprade 4.y.z -> 4.y+1.z' might not be completed
up to a bug in 4.y+1.z' that has a fix in 4.(y+1).z''.
We need the retarget to it to land 4.y+1 on the cluster.

* Generate accepted risks in an unblocking case

This commit generates a message in the accepted risks
for the unblocking case 4.y.z -> 4.y+1.z' -> 4.y+1.z''.

* Use Progressing=True to block minor level upgrade

* Drop the guard on UpgradeInProgress

The guard will be added on another pull request.

* Make the type UpgradeableUpgradeInProgress private

* Improve condition's reason and message