pkg: Support ClusterVersionOperator API
Create a feature gated controller to watch for changes on the ClusterVersionOperator
`cluster` resource. On creation or modification, run a synchronization. Put relevant logic
behind the CVOConfiguration feature gate to limit the new DevPreviewNoUpgrade logic from
the Default logic as much as possible.

The purpose of the `Start` method is to make the creation of a ClusterVersionOperator informer
optional, and to make the creation possible later in the CVO logic by restarting the operator
informer factory. To decide whether to create the informer, feature gates must be known,
which happens later in the run when all the other informer factories are already created an
inaccessible to the main CVO controller.

Related enhancement: [1]

[1]: https://github.com/openshift/enhancements/blob/2890cccf20ebcb94fce901f7afb170ca680aa2d9/enhancements/update/cvo-log-level-api.md