OTA-1418: Allow forcing health insights on CV
When a `ClusterVersion` is annotated with `usc.openshift.io/force-health-insight`, make USC emit a health insight, suitable for testing health insight lifecycle. Note that such health insight never goes away for now; this will be addressed by followup PRs.

Health insights UIDs are a hash of the insight message and the resources involved. This will become problematic if we ever want to do aggregation, but for now it suits the purpose.