OTA-1418: Identify the informer sending insight message
Make USC aware of the identities of the informers that send insights to it, by including the identifier in the message sent to USC. In the future API, we will have separated containers per informer:

```
informers:
- name: name-1
  insights: [...]
- name: name-2
  insights: [...]
```

We do not have this in our ConfigMap API mock so for now we encode the informer into the key of the item stored in the ConfigMap. USC also takes ownership of the `usc.` prefix (formerly `usc-`) which identifies insight keys, and the informers no longer need to maintain this contract.