USC: Drop unknown messages from the API
Assume the informers, in each message, include a list of all active insight IDs that infomer knows about. The Status API can then drop the unknown insights formerly reported by that informer (=insights present in the API but not known to be active by the responsible informer)

This is only a part of OTA-1418. The remaining changes will be as follows:

- Do not drop the unknown insights right away but make them expire after some some time of not being renewed (handle the case where informer starts up and starts sending messages about insights as it discovers them)
- Actually make the informers remember their insights and populate the `knownInsights` field of the message
- Add additional `deletedInsights` field to the message, to allow informers explicitly tell the API some insight went away
- Make the USC respect `deletedInsights`
- Make informers fill `deletedInsight`
- Potentially add a out-of-API memory cache for insights to be expired / deleted but allow them to reappear with formerly known content

Co-authored-by: David Hurta <dhurta@redhat.com>