OTA-1418: USC: Make messages more robust and decoupled (#1146)
* OTA-1418: Identify the informer sending insight message

Make USC aware of the identities of the informers that send insights to it, by including the identifier in the message sent to USC. In the future API, we will have separated containers per informer:

```
informers:
- name: name-1
  insights: [...]
- name: name-2
  insights: [...]
```

We do not have this in our ConfigMap API mock so for now we encode the informer into the key of the item stored in the ConfigMap. USC also takes ownership of the `usc.` prefix (formerly `usc-`) which identifies insight keys, and the informers no longer need to maintain this contract.

* USC: Drop & log invalid messages

Co-authored-by: David Hurta <dhurta@redhat.com>

* USC: Validate insight message creation in producers

Co-authored-by: David Hurta <dhurta@redhat.com>

* USC: Producers no longer need to prefix insight UIDs

Co-authored-by: David Hurta <dhurta@redhat.com>

* USC: Address review

---------

Co-authored-by: David Hurta <dhurta@redhat.com>