`status`: extract API types to a package
We are temporarily using API types that will eventually come via importing openshift/api => under a package. Extract the current API types to a subpackage, so everything we use from it is used through a package (e.g. `updatestatus.THING` instead of just `THING`). This will make an eventual switch to the proper API easier.