OTA-1418: Decribe informer->usc communication
Going from USC-scrapes-all-insights periodically to unidirectional
message passing makes insight lifecycle a bit harder (if an informer is
not saying anyting about an insight Status API knows about, is it just
unchanged, or did the informer lose track of it?).

List of known insight UIDs is lightweight-enough payload to be included
in every message sent to USC and the USC can (eventually) drop the
insights that the informer stops reporting as known, and can drop
the insights the informer requests to be dropped.

Unsure whether this mechanism survives until the final version but it is
reasonably robust for the current state.