pkg: Add CVOConfiguration controller
The controller has an empty reconcile logic as of now.
The logic will be implemented later. The goal of this commit is to introduce
a new CVO controller, which can optionally (depending on the state of the
CVOConfiguration feature gate) create a new informer.

The purpose of the `Start` method is to make the creation of a ClusterVersionOperator informer
optional, and to make the creation possible later in the CVO logic by restarting the operator
informer factory. To decide whether to create the informer, feature gates must be known,
which happens normally later in the run when all the other informer factories are already
created and inaccessible to the main CVO controller.

Related enhancement: [1]

[1]: https://github.com/openshift/enhancements/blob/2890cccf20ebcb94fce901f7afb170ca680aa2d9/enhancements/update/cvo-log-level-api.md