pkg/internal: Add log level constants
These constants can be simply used when logging across the project instead
of specifying the level by a number. The descriptions and the values will
also guide developers when deciding what log level to choose.

For example:

```go
  import "github.com/openshift/cluster-version-operator/pkg/internal"

  klog.V(internal.Debug).Infof("")
```

The descriptions of the values are extracted from the used API type [1].

[1]: https://github.com/openshift/api/blob/744790f2cff777b1bb29bdccce10e4e84cff0a69/operator/v1/types.go#L94-L109