Move sync.Once into nodeInformerController controller level
As a result each nodeInformerController instance will execute
the function once which makes more sense than doing it once
grobally because what we do there is initialization of the
caches that are associated with the instance.

For the core code, it is not a big deal since we have only ONE
instance of the controller. However, it matters for unit tests where
there are many controllers.