Reconcile all nodes via a special event
This is to improve the error handling on `reconcileAllNodes()`.
See [1] for details.

Prior to this commit, the event on MC/MCP will be re-queued if `reconcileAllNodes()`
hits an error. However, `syncMachineConfig()` (or `syncMachineConfigPool` respectively)
is stateful, i.e., the result replies on the content of the caches that might be
changed from the original event.

With the commit, a special event stays between an MC/MCP event and `reconcileAllNodes()`.
An error from the latter will re-queue the special event which basically means triggering
another run of `reconcileAllNodes()`.

[1]. https://github.com/openshift/cluster-version-operator/pull/1144#discussion_r1969935838