cvo: Add `config-file` flag
Add a new flag instead of a new environment variable to specify the path of the
configuration file. Using env variables has the benefit of better backwards
compatibility, for example, specifying not used env variable has no impact. Using
a not-implemented flag results in an error. The flag is primarily intended for use
in HyperShift, thus, these topics are relevant.

In case the feature gate does not get into default, we must not drop the flag
out of nowhere, which could cause issues in HyperShift. However, the new API
is very small, and it is not likely for this functionality to be dropped
along the way. An alternative is to use an env variable before the feature gets
into default; however, I do not think this is worth the risk at the moment.